refactor: move node dimension logic to saga-flow-sizing to reduce component file size
Moves getNodeDimensions and DEFAULT_NODE_DIMENSIONS to the sizing
utility file. Reduces saga-flow.tsx from 634 to 611 lines (under
the 614 allowlist limit).
